+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,12 +1,15 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain;
+using Restaurants.Domain.Constants;
+using Restaurants.Domain.Interfaces;
 
 namespace Restaurants.Application
 {
     public class DeleteRestaurantCommandHandler(
         ILogger<DeleteRestaurantCommandHandler> logger,
-        IRestaurantRepository restaurantRepository) : IRequestHandler<DeleteRestaurantCommand>
+        IRestaurantRepository restaurantRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -14,6 +17,9 @@ namespace Restaurants.Application
             var restaurant = await restaurantRepository.GetByIdAsync( request.Id );
             if (restaurant == null)
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, RestaurantOperations.Delete))
+                throw new ForbidException();
 
             await restaurantRepository.DeleteAsync(restaurant);
         }
